% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summary.R
\name{my_summary}
\alias{my_summary}
\title{Linear Model Summary}
\usage{
my_summary(object)
}
\arguments{
\item{object}{An object returned by my_lm function.}
}
\value{
multiple vectors of coefficients, tests, and R square values.
}
\description{
Summarize linear model to the given data.
}
\examples{
my_summary(model)

}
