% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mylm.R
\name{print.mylm}
\alias{print.mylm}
\title{Print Method for My Linear Model Objects}
\usage{
\method{print}{mylm}(x, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}

\item{model}{An model returned by my_lm function.}
}
\value{
A numeric vector of coefficients.
}
\description{
Print the result of a linear model.
}
\examples{
print.mylm(my_lm(mpg ~ wt + cyl, data = mtcars))

}
